@@ -3,14 +3,19 @@
   "version": "1.5.1",
   "description": "A Vue digital clock component, for use out the box",
   "author": "Edd Yerburgh <edward.yerburgh@gmail.com>",
-  "main": "./Clock.vue",
+  "main": "./dist/cjs/Clock.js",
+  "module": "./dist/es/Clock.js",
+  "files": [
+    "dist/**/*.js"
+  ],
   "scripts": {
-    "dev": "node build/dev-server.js",
-    "build": "node build/build.js",
-    "unit": "karma start test/unit/karma.conf.js --single-run",
-    "release": "semantic-release",
-    "test": "npm run unit",
-    "lint:fix": "eslint --ext .js,.vue Clock.vue test/unit/specs --fix"
+    "build": "rollup -c",
+    "dev": "parcel example/index.html",
+    "format": "prettier --write \"{,!(.cache|dist)/**/}/*.{md,vue,js,json}\"",
+    "format:check": "prettier --check \"{,!(.cache|dist)/**/}/*.{md,vue,js,json}\"",
+    "prepare": "npm run build",
+    "test:unit": "jest",
+    "test": "npm run format:check && npm run test:unit"
   },
   "keywords": [
     "vue clock",
@@ -35,59 +40,27 @@
     "vue js digital clock component"
   ],
   "peerDependencies": {
-    "vue": "^2.1.0"
+    "vue": "^2.x"
   },
   "devDependencies": {
+    "@babel/core": "^7.2.2",
+    "@babel/preset-env": "^7.3.1",
     "avoriaz": "^1.8.2",
-    "chai": "^3.5.0",
-    "chalk": "^1.1.3",
-    "connect-history-api-fallback": "^1.1.0",
-    "cross-env": "^3.1.4",
-    "css-loader": "^0.25.0",
-    "eslint": "^3.7.1",
-    "eslint-config-airbnb-base": "^8.0.0",
-    "eslint-friendly-formatter": "^2.0.5",
-    "eslint-import-resolver-webpack": "^0.6.0",
-    "eslint-loader": "^1.5.0",
-    "eslint-plugin-html": "^1.3.0",
-    "eslint-plugin-import": "^1.16.0",
-    "eventsource-polyfill": "^0.9.6",
-    "express": "^4.13.3",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "friendly-errors-webpack-plugin": "^1.1.2",
-    "function-bind": "^1.0.2",
-    "html-webpack-plugin": "^2.8.1",
-    "http-proxy-middleware": "^0.17.2",
-    "inject-loader": "^2.0.1",
-    "json-loader": "^0.5.4",
-    "karma": "^1.3.0",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-coverage": "^1.1.1",
-    "karma-mocha": "^1.2.0",
-    "karma-phantomjs-launcher": "^1.0.0",
-    "karma-sinon-chai": "^1.2.0",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-spec-reporter": "0.0.26",
-    "karma-webpack": "^1.7.0",
-    "lolex": "^1.4.0",
-    "mocha": "^3.1.0",
-    "opn": "^4.0.2",
-    "ora": "^0.3.0",
-    "phantomjs-prebuilt": "^2.1.3",
-    "semantic-release": "^15.9.17",
-    "semver": "^5.3.0",
-    "shelljs": "^0.7.4",
-    "sinon": "^2.1.0",
-    "sinon-chai": "^2.9.0",
-    "url-loader": "^0.5.7",
-    "vue-loader": "^10.0.0",
-    "vue-style-loader": "^1.0.0",
-    "vue-template-compiler": "^2.1.0",
-    "webpack": "^1.13.2",
-    "webpack-dev-middleware": "^1.8.3",
-    "webpack-hot-middleware": "^2.12.2",
-    "webpack-merge": "^0.14.1"
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-jest": "^24.0.0",
+    "jest": "^24.0.0",
+    "parcel": "^1.11.0",
+    "prettier": "^1.16.3",
+    "rollup": "^1.1.2",
+    "rollup-plugin-buble": "^0.19.6",
+    "rollup-plugin-commonjs": "^9.2.0",
+    "rollup-plugin-node-resolve": "^4.0.0",
+    "rollup-plugin-vue": "^4.6.2",
+    "sinon": "^7.2.3",
+    "vue": "^2.5.22",
+    "vue-hot-reload-api": "^2.3.1",
+    "vue-jest": "^3.0.2",
+    "vue-template-compiler": "^2.5.22"
   },
   "engines": {
     "node": ">= 4.0.0",
@@ -100,5 +73,11 @@
   "bugs": {
     "url": "https://github.com/eddyerburgh/vue-digital-clock/issues"
   },
-  "homepage": "https://github.com/eddyerburgh/vue-digital-clock#readme"
+  "homepage": "https://github.com/eddyerburgh/vue-digital-clock#readme",
+  "jest": {
+    "transform": {
+      "^.+\\.js$": "babel-jest",
+      "^.+\\.vue$": "vue-jest"
+    }
+  }
 }
