@@ -1,11 +1,9 @@
-// The Vue build version to load with the `import` command
-// (runtime-only or standalone) has been set in webpack.base.conf with an alias.
-const Vue = require('vue');
-const App = require('./App');
+import Vue from "vue";
+import App from "./App.vue";
 
 /* eslint-disable no-new */
 new Vue({
-  el: '#app',
-  template: '<App/>',
-  components: { App },
+  el: "#app",
+  render: h => h(App),
+  components: { App }
 });
